pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pupper-can-adapter"
version = "0.1.0"
description = "Actuation Abstraction Layer to run Stanford Pupper gaits on CAN-based actuators"
authors = [{name = "just0P"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-can>=4.6",
  "PyYAML>=6.0",
  "numpy>=1.26",
]

[project.scripts]
pupper-can-scan = "pupper_can_adapter.scripts.can_scan:main"
pupper-can-run = "pupper_can_adapter.scripts.run_can_walk:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["pupper_can_adapter*"]
